boot: x86 bootsector hello world
This initial commit adds initial README, license, and an x86 bootsector
implementation of a 'hello, world' program.

This is to provide an initial reference point from which to iterate the
rest of the kernel implementation and to get minimal working code running.

The Makefile adds a convenient 'make qemu' option to build and run in
qemu immediately, which will of course be the core development environment.

As we are using initial realmode we can use the x86-64 qemu, as all x86
chips start in 16-bit real mode. We eventually plan to target x86-64 so it
makes sense to do this from the outset.