boot: load kernel ELF image and run it
Due to code size restrictions in the bootsector we have to seriously cut
down what we're doing and perform far fewer checks. This should be
updated with some code to load at least another disk sector's worth of
data to avoid the need for 'squashing' things up.

We load an ELF image and execute it, finally reaching a state of having
a 64-bit long, protected, page kernel mode executable functioning with
1 GiB of virtual address space mapped.

As 'proof of life' the kernel main() function simply outputs 'Hello'
to the video output.