boot: clean up boot assembly, use more defines for values, output more errors
Now that we don't need to squeeze the full boot loader into 512 bytes we can
improve and clean up the boot assembly, outputting more meaningful errors if
things go wrong, cleaning up the code and using defines where it makes sense
to do so.

We also update what we use as the initial kernel stack pointer, the top of
conventional memory which provides plenty of space for kernel operations.