mm: read e820 physical memory layout attributes
The %eax=0xe820 int 0x16 call provides a full physical memory map for x86
systems and obtaining it is vital to being able to know what memory is
available to use.

Add code to retrieve this data during real mode when we can still access
it, place it in a memory region that the kernel can later access.

The patch also dumps the output on OS start.