test: add userland test runner
Some things can be tested in userland and make more sense to be tested
there so create a test runner for this.

The concept is that headers and source files (where appropriate!) can be
imported and tested in an unconstrained environment.

The motivating initial use case is a spinlock implementation which in
userland outside of the kernel entirely can be run in mulitple
std::thread's.

We additionally add a general 'make test' option to run both early and
userland based tests together.