spinlock: add initial implementation
The fundamentals of a spinlock are busy-waiting on a atomic exchange.
This refined version is inspired by the excellent, referenced article
and uses compiler intrinsic atomic operations with memory model specified
and CPU hinting to improve performance.

We add this now as a prerequisite to the physical allocator implementation.
While we have not yet implemented anything SMP, the physical allocator will
absolutely need to acquire locks so having them ready to go from now is the
right way to go.

The userland testing framework can be used to test them using std::thread's
and do things that would be very difficult with this extremely early stage
kernel.