mm: take a copy of a physical block when accesssing
Since a physical block is very small (16 bytes) and to avoid making
accessors with locks for every field, spinlock a block and obtain the
full struct rather than working with a pointer to it.

We will need to revisit how to efficiently store information about
physical memory pages at a later point.