build: Eliminate unnecessary no-pic, no-pie specifiers and -mcmodel=large
These don't appear to be necessary, we can perform relative jumps without issues.

This also resolves a hard to diagnose issue with the -fgcse optimisation which
was causing a boot cycle. It is uncertain as to how! Perhaps the compiler thought
separate values were in fact the same. It may be something we need to return to.