early: mem: provide scratch page allocator
Provide a SUPER early physical page allocator that simply allocates pages
past where we place the ELF image. None of these pages are intended to
be kept and we provide no freeing mechanism.

We need this in order to allocate pages for our early page allocator. We
have this in place in order to avoid a chicken and egg situation with
the early page allocator initialisation sequence.