zbuild: initial incomplete version that can parse
zbuild is conceived to be a more logical means of building zeptux than a
Makefile or cmake (ugh) that can be customised according to the needs of
the project.

In this patch I introduce an imagined zbuild file which generally matches
the existing Makefile (but more succinct and correct!) and some first
draft code that parses this file.

I am using go as a fast 'batteries included' environment for this work.