build: Move to using zbuild exclusively to build zeptux
Keeping the Makefile updated has become a huge pain and cmake is a horror so I
rolled my own build system in a somewhat rushed and less than perfect go
implementation.

This may or may not turn out to be a dreadful mistake but it is a hill I am
willing to die on so let's go all in and move to using zbuild exclusively.