zbuild: add global_file_deps special variable
Note that this doesn't impact foreach expansions however this is fine -
we absolutely do not need to expand foreach's except when the target file
is newer than the source. All rules otherwise will get rebuild when the
file(s) are changed.

This is specifically designed to allow us to force rebuilds when the
zeptux.zbuild file itself is changed.