build: separate out kernel/main.c build object
We want to put more C files in the kernel/ directory that need to be
imported to early test runs. Take advantage of new zbuild functionality
to be able to exclude paths from a build statement in order to separate
out kernel/main.c correctly while still being able to add further C files
in the kernel/ directory.