early: mm: add skeleton early direct memory mapping implementation
Add the logic in the early memory code to invoke the direct memory map
and skeleton code to perform general memory maps of ranges which this
invokes.

This helps us get our hands around the scope of the task as well as
implementing the early memory portion of the code.