page: add p*d_to_virt_ptr(), assign_p*d() helpers
The p*d_to_virt_ptr() helpers simply obtain a pointer to the specified
page table via the direct mapping, assign_p*d() assigns the specified
page table address to the page table entry in the level above it, e.g.
assign_pud() sets an entry in the PGD at the specified index to the
PUD being assigned.

The assign_p*d() entries all assign higher level pages (and in the case
of PUD and PMD are specifically for the 4 KiB page size) so we do not
need to do much in terms of flags - setting present and R/W is sufficient.

The specific flags will be set in the PTDE or in the case of 2 MiB, 1 GiB
pages in the PMDE and PUDE respectively which will be implemented in
separate functions.