test: early: mm: we cannot always perform the crossing span tests
As we use more lower memory, this isn't always possible. We probably need
to find a way around this but for the time being disable the 'next span' test
in this situation. If memory is set >= 8G in qemu then the test runs again so
this is a temporary albeit dirty workaround.