early: mm: relocate the early scratch patches prior to the ELF image
This is to avoid the issue where static variables are placed immediately
after the ELF image loaded from the file, and the scratch allocator could
actually allocate a page including the scratch allocator state static
variable (!) resulting in the scratch alloc state becoming corrupted,
most noticeably with the start address being zeroed meaning that
page tables got overwritten resulting in page faults.

This is not a good type of heisenbug to deal with, so we relocate this
memory to avoid it.