build: elf: offset .bss to avoid overlap between section heads + it
The section headers are located at the end of the ELF image but can be located
such that they overlap with the virtual address allocated to the .bss, given
that the ELF is loaded as one contiguous mass of data.

This is a product of loading in a 'dumb' fashion, a memcpy() more or less.

This results in memory corruption as static memory overwrites section headers
and the kernel ELF section mapping code fails.

Workaround this by offsetting the .bss by several pages to ensure no overlap
can reasonably occur.